requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5